Turn distance respawn cooldown into respawn time
Having it be a cooldown created this weird buffering effect, where the first
respawn would be instantaneous but then the cooldown would kick in for the next
one. Now there is a set time between moving the vehicle away from its spawn
position and the replacement appearing.

Vehicles are also added to zeus as soon as they leave their respawn distance,
not just when the replacement has spawned.